\name{HeatmapPlot-package}
\alias{HeatmapPlot-package}
\alias{HeatmapPlot}
\title{Longitudinal Omics Heatmaps and Temporal Module Plots}
\description{
Cluster longitudinal omics features into temporal modules, draw ordered
ComplexHeatmap heatmaps, summarize module trajectories, and align paired omics
layers such as RNA and ATAC by gene.
}
\details{
Core workflow functions are \code{longitudinal_cluster()},
\code{longitudinal_heatmap()}, \code{pattern_lineplot()}, and
\code{write_memberships()}. Public real-data examples can be loaded with
\code{load_real_timecourse_example()}.
}
\keyword{package}
